requirements.txt
Django==1.5.1
psycopg2==2.5
wsgiref==0.1.2
django-user-accounts
